Changed a int to a float.
git-svn-id: https://lexevs.svn.sourceforge.net/svnroot/lexevs/base/v6/trunk/lexevs_api@8865 c4b421de-a31f-834e-8902-8020e42e311c

+++ cts2Impl/src/main/java/org/cts2/internal/match/LevenshteinDistanceMatcher.java
@@ -46,7 +46,7 @@ public class LevenshteinDistanceMatcher extends AbstractMatcher {
 	 * @param longestString the longest string
 	 * @return the float
 	 */
-	protected float normalize(int score, int longestString){
-		return 1 - (score / longestString);
+	protected float normalize(float score, int longestString){
+		return 1f - (score / longestString);
 	}
 }
